chore: drop hand-written DnD schemas
