+++ dnd/schemas/npc.schema.json
@@ -1,20 +1,99 @@
 {
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "NPC",
-  "type": "object",
-  "required": ["id", "name", "tags"],
-  "properties": {
-    "id": { "type": "string" },
-    "name": { "type": "string" },
-    "tags": {
-      "type": "array",
-      "items": { "type": "string" }
-    },
-    "description": { "type": "string" },
-    "stats": {
+  "$ref": "#/definitions/npc",
+  "definitions": {
+    "npc": {
       "type": "object",
-      "additionalProperties": { "type": "number" }
+      "properties": {
+        "id": {
+          "type": "string",
+          "minLength": 1
+        },
+        "name": {
+          "type": "string",
+          "minLength": 1
+        },
+        "species": {
+          "type": "string",
+          "minLength": 1
+        },
+        "role": {
+          "type": "string",
+          "minLength": 1
+        },
+        "alignment": {
+          "type": "string",
+          "minLength": 1
+        },
+        "backstory": {
+          "type": "string"
+        },
+        "location": {
+          "type": "string"
+        },
+        "hooks": {
+          "type": "array",
+          "items": {
+            "type": "string"
+          },
+          "minItems": 1
+        },
+        "quirks": {
+          "type": "array",
+          "items": {
+            "type": "string"
+          }
+        },
+        "voice": {
+          "type": "object",
+          "properties": {
+            "style": {
+              "type": "string",
+              "minLength": 1
+            },
+            "provider": {
+              "type": "string",
+              "minLength": 1
+            },
+            "preset": {
+              "type": "string",
+              "minLength": 1
+            }
+          },
+          "required": [
+            "style",
+            "provider",
+            "preset"
+          ],
+          "additionalProperties": false
+        },
+        "portrait": {
+          "type": "string"
+        },
+        "statblock": {
+          "type": "object",
+          "additionalProperties": {}
+        },
+        "tags": {
+          "type": "array",
+          "items": {
+            "type": "string"
+          },
+          "minItems": 1
+        }
+      },
+      "required": [
+        "id",
+        "name",
+        "species",
+        "role",
+        "alignment",
+        "hooks",
+        "voice",
+        "statblock",
+        "tags"
+      ],
+      "additionalProperties": false
     }
   },
-  "additionalProperties": false
+  "$schema": "http://json-schema.org/draft-07/schema#"
 }
