+++ scripts/gen-schemas.ts
@@ -1,0 +1,18 @@
+import { promises as fs } from "fs";
+import path from "path";
+import { zodToJsonSchema } from "zod-to-json-schema";
+import { zNpc } from "../src/dnd/schemas/npc";
+import { zLore } from "../src/features/dnd/schemas";
+
+async function main() {
+  const outDir = path.resolve("dnd", "schemas");
+  await fs.mkdir(outDir, { recursive: true });
+
+  const npcJson = zodToJsonSchema(zNpc, "npc");
+  await fs.writeFile(path.join(outDir, "npc.schema.json"), JSON.stringify(npcJson, null, 2));
+
+  const loreJson = zodToJsonSchema(zLore, "lore");
+  await fs.writeFile(path.join(outDir, "lore.schema.json"), JSON.stringify(loreJson, null, 2));
+}
+
+main();
