+++ src/dnd/schemas/npc.ts
@@ -1,11 +1,25 @@
 import { z } from "zod";
 
-export const npcSchema = z.object({
-  id: z.string(),
-  name: z.string(),
-  tags: z.array(z.string()),
-  description: z.string().optional(),
-  stats: z.record(z.number()).optional()
+export const zVoice = z.object({
+  style: z.string().min(1),
+  provider: z.string().min(1),
+  preset: z.string().min(1),
 });
 
-export type Npc = z.infer<typeof npcSchema>;
+export const zNpc = z.object({
+  id: z.string().min(1),
+  name: z.string().min(1),
+  species: z.string().min(1),
+  role: z.string().min(1),
+  alignment: z.string().min(1),
+  backstory: z.string().optional(),
+  location: z.string().optional(),
+  hooks: z.array(z.string()).nonempty(),
+  quirks: z.array(z.string()).optional(),
+  voice: zVoice,
+  portrait: z.string().optional(),
+  statblock: z.record(z.unknown()),
+  tags: z.array(z.string()).nonempty(),
+});
+
+export type Npc = z.infer<typeof zNpc>;
