--- src/features/dnd/lore.schema.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "type": "object",
-  "required": ["id", "name", "summary", "tags"],
-  "properties": {
-    "id": { "type": "string", "minLength": 1 },
-    "name": { "type": "string", "minLength": 1 },
-    "summary": { "type": "string", "minLength": 1 },
-    "location": { "type": "string" },
-    "hooks": { "type": "array", "items": { "type": "string" } },
-    "tags": { "type": "array", "items": { "type": "string" }, "minItems": 1 }
-  }
-}
