--- src/features/dnd/npc.schema.json
@@ -1,28 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "type": "object",
-  "required": ["id", "name", "species", "role", "alignment", "hooks", "voice", "statblock", "tags"],
-  "properties": {
-    "id": { "type": "string", "minLength": 1 },
-    "name": { "type": "string", "minLength": 1 },
-    "species": { "type": "string", "minLength": 1 },
-    "role": { "type": "string", "minLength": 1 },
-    "alignment": { "type": "string", "minLength": 1 },
-    "backstory": { "type": "string" },
-    "location": { "type": "string" },
-    "hooks": { "type": "array", "items": { "type": "string" }, "minItems": 1 },
-    "quirks": { "type": "array", "items": { "type": "string" } },
-    "voice": {
-      "type": "object",
-      "properties": {
-        "style": { "type": "string", "minLength": 1 },
-        "provider": { "type": "string", "minLength": 1 },
-        "preset": { "type": "string", "minLength": 1 }
-      },
-      "required": ["style", "provider", "preset"]
-    },
-    "portrait": { "type": "string" },
-    "statblock": { "type": "object" },
-    "tags": { "type": "array", "items": { "type": "string" }, "minItems": 1 }
-  }
-}
