+++ src/features/dnd/schemas.ts
@@ -1,4 +1,5 @@
 import { z } from "zod";
+import { zNpc } from "../../dnd/schemas/npc";
 
 export const zDndBase = z.object({
   id: z.string().min(1),
@@ -11,27 +12,7 @@ export const zDndTheme = z.enum(["Parchment", "Ink", "Minimal"]);
  * Minimal schema for non-player characters (NPCs).
  * Designed to keep downstream consumers consistent and predictable.
  */
-export const zVoice = z.object({
-  style: z.string().min(1),
-  provider: z.string().min(1),
-  preset: z.string().min(1),
-});
-
-export const zNpc = z.object({
-  id: z.string().min(1),
-  name: z.string().min(1),
-  species: z.string().min(1),
-  role: z.string().min(1),
-  alignment: z.string().min(1),
-  backstory: z.string().optional(),
-  location: z.string().optional(),
-  hooks: z.array(z.string()).nonempty(),
-  quirks: z.array(z.string()).optional(),
-  voice: zVoice,
-  portrait: z.string().optional(),
-  statblock: z.record(z.unknown()),
-  tags: z.array(z.string()).nonempty(),
-});
+export { zNpc };
 
 /**
  * Simple schema for world lore entries.
